+++ src/main/java/racingcar/RacingCars.java
@@ -2,6 +2,7 @@ package racingcar;
 
 import exception.CustomException;
 import racingcar.exception.RacingCarErrorCode;
+import racingcar.generator.NumberGenerator;
 
 import java.util.ArrayList;
 import java.util.List;
