+++ src/main/java/racingcar/generator/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.generator;
 
 public interface NumberGenerator {
